% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_basic_recursion.R
\name{ackermann}
\alias{ackermann}
\title{Standard functions implemented by recursion}
\usage{
ackermann(m, n)
}
\arguments{
\item{m}{A positive integer.}

\item{n}{A positive integer.}
}
\description{
Standard functions implemented by recursion
}
\details{
Functions leveraging recursion.

This function will only finish in a reasonable amount of time
for m and n less than or equal to 4.
}
\examples{
library(functionalPlayground)

ackermann(2, 1)
ackermann(1, 2)
}
