% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_function_operators.R
\name{compose}
\alias{compose}
\title{Standard function operator.}
\usage{
compose(f, g)
}
\arguments{
\item{f}{A function.}

\item{g}{A function.}
}
\description{
Standard function operator.
}
\details{
Mathematical operation compose. (fog)(x) = f(g(x))
}
\examples{
library(functionalPlayground)

num_even <- compose(recursive_add, is_even)
num_even(1:10)
}
