% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_higher_order.R
\name{foldl}
\alias{foldl}
\title{Higher order functions}
\usage{
foldl(f, id)
}
\arguments{
\item{f}{A function.}

\item{id}{Value when the input vector is empty}
}
\description{
Higher order functions
}
\details{
The returned function takes a single argument \code{x}, a vector that is
numeric or logical depending on \code{f}.
}
\keyword{internal}
