% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_mutual_recursion.R
\name{hofstadter_female}
\alias{hofstadter_female}
\alias{hofstadter_male}
\title{Mutual recursion}
\usage{
hofstadter_female(n)

hofstadter_male(n)
}
\arguments{
\item{n}{A numeric vector containing one integer.}
}
\description{
Mutual recursion
}
\details{
Functions leveraging mutual recursion.
}
\examples{
library(functionalPlayground)

n <- 3
hofstadter_female(n)
hofstadter_male(n)
}
