% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_basic_recursion.R
\name{hofstadter_g}
\alias{hofstadter_g}
\title{Standard functions implemented by recursion}
\usage{
hofstadter_g(n)
}
\arguments{
\item{n}{A numeric vector containing one integer.}
}
\description{
Standard functions implemented by recursion
}
\details{
Functions leveraging recursion.
}
\examples{
library(functionalPlayground)

hofstadter_g(0)
hofstadter_g(1)
hofstadter_g(2)
hofstadter_g(3)
}
