% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_function_operators.R
\name{negate}
\alias{negate}
\title{Standard function operator.}
\usage{
negate(f)
}
\arguments{
\item{f}{A function to be negated.}
}
\description{
Standard function operator.
}
\details{
Negate any function.
}
\examples{
library(functionalPlayground)

always_true <- function(...) {
  return(TRUE)
}
always_false <- negate(always_true)
always_false(TRUE)
always_false(FALSE)
}
