% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_basic_recursion.R
\name{recursive_gcd}
\alias{recursive_gcd}
\alias{recursive_lcm}
\title{Standard functions implemented by recursion}
\usage{
recursive_gcd(a, b)

recursive_lcm(a, b)
}
\arguments{
\item{a}{A positive integer vector of length 1.}

\item{b}{A positive integer vector of length 1.}
}
\description{
Standard functions implemented by recursion
}
\details{
Functions leveraging recursion.
}
\examples{
library(functionalPlayground)

recursive_gcd(4L, 6L)
recursive_gcd(10L, 20L)
recursive_gcd(1L, 100L)

recursive_lcm(4L, 6L)
recursive_lcm(10L, 20L)
recursive_lcm(1L, 100L)
}
