% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_basic_recursion.R
\name{recursive_length}
\alias{recursive_length}
\title{Standard functions implemented by recursion}
\usage{
recursive_length(x)
}
\arguments{
\item{x}{A vector that is not empty.}
}
\description{
Standard functions implemented by recursion
}
\details{
Functions leveraging recursion.
}
\examples{
library(functionalPlayground)

recursive_length(1:10)
recursive_length(1)
}
