% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2 base functions.R
\name{add}
\alias{add}
\title{Simple helper functions.}
\usage{
add(x1, x2)
}
\arguments{
\item{x1}{A vector of length 1. Numeric or logical depending on function.}

\item{x2}{A vector of length 1. Numeric or logical depending on function.}
}
\description{
Simple helper functions.
}
\keyword{internal}
