% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_cumulative_numeric_functions.R
\name{cumulative_add}
\alias{cumulative_add}
\alias{cumulative_multiply}
\alias{cumulative_max}
\alias{cumulative_min}
\title{Cumulative functions implemented in a functional style}
\usage{
cumulative_add(x)

cumulative_multiply(x)

cumulative_max(x)

cumulative_min(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Cumulative functions implemented in a functional style
}
\details{
Functions leveraging map and a function operator.
}
\examples{
library(functionalPlayground)

x <- 1:10
cumulative_add(x)
cumulative_multiply(x)
cumulative_max(x)
cumulative_min(x)
}
