% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_cumulative_logical_functions.R,
%   R/8_function_operators.R
\name{cumulative_all}
\alias{cumulative_all}
\alias{cumulative_any}
\alias{cumulative_none}
\title{Cumulative functions implemented in a functional style}
\usage{
cumulative_all(x)

cumulative_any(x)

cumulative_none(x)
}
\arguments{
\item{x}{A logical vector.}
}
\description{
Cumulative functions implemented in a functional style
}
\details{
Functions leveraging map and a function operator.
}
\examples{
library(functionalPlayground)

x <- c(TRUE, TRUE, FALSE, FALSE)
cumulative_all(x)
cumulative_any(x)
cumulative_none(x)
}
