% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6 cumulative logical functions.R
\name{cumulative_all}
\alias{cumulative_all}
\alias{cumulative_any}
\title{Standard functions implemented in a functional style}
\usage{
cumulative_all(x)

cumulative_any(x)
}
\arguments{
\item{x}{A logical vector.}
}
\description{
Standard functions implemented in a functional style
}
\details{
Functions leverage map and a function operator.
}
\examples{
library(functionalPlayground)

x <- c(TRUE, TRUE, FALSE, FALSE)
cumulative_all(x)
cumulative_any(x)
}
