% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_higher_order_functions.R
\name{foldl}
\alias{foldl}
\title{Higher order functions}
\usage{
foldl(f, id)
}
\arguments{
\item{f}{A function.}

\item{id}{Value when x is empty}

\item{x}{A vector. Numeric or logical depending on function.}
}
\description{
Higher order functions
}
\keyword{internal}
