% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_basic_recursive_functions.R
\name{hofstadter_h}
\alias{hofstadter_h}
\title{Standard functions implemented by recursion}
\usage{
hofstadter_h(n)
}
\arguments{
\item{n}{A numeric vector containing one integer.}
}
\description{
Standard functions implemented by recursion
}
\details{
Functions leveraging recursion.
}
\examples{
library(functionalPlayground)

hofstadter_h(0)
hofstadter_h(1)
hofstadter_h(2)
hofstadter_h(3)
}
