% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_basic_recursive_functions.R
\name{hofstadter_q}
\alias{hofstadter_q}
\title{Standard functions implemented by recursion}
\usage{
hofstadter_q(n)
}
\arguments{
\item{n}{A numeric vector containing one integer.}
}
\description{
Standard functions implemented by recursion
}
\details{
Functions leveraging recursion.
}
\examples{
library(functionalPlayground)

hofstadter_q(1)
hofstadter_q(2)
hofstadter_q(3)
}
