% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7 mutual recursion.R
\name{is.even}
\alias{is.even}
\alias{is.odd}
\title{Standard functions implemented in a functional style}
\usage{
is.even(x)

is.odd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Standard functions implemented in a functional style
}
\details{
Functions leverage mutual recursion.
}
\examples{
library(functionalPlayground)

x <- 1:10
is.even(x)
is.odd(x)
}
