% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_mutual_recursion.R
\name{is_even}
\alias{is_even}
\alias{is_odd}
\title{Mutual recursion}
\usage{
is_even(x)

is_odd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Mutual recursion
}
\details{
Functions leveraging mutual recursion.
}
\examples{
library(functionalPlayground)

x <- 1:10
is_even(x)
is_odd(x)
}
