% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8 negate.R
\name{negate}
\alias{negate}
\title{Standard functions implemented in a functional style}
\usage{
negate(f)
}
\arguments{
\item{f}{A function to be negated.}
}
\description{
Standard functions implemented in a functional style
}
\details{
Negate any function
}
\examples{
library(functionalPlayground)

always_true <- function(...) {
return(TRUE)
}
always_false <- negate(always_true)
always_false(TRUE)
always-false(FALSE)
}
