% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3 fold numeric functions.R
\name{recursive_add}
\alias{recursive_add}
\alias{recursive_multiply}
\alias{recursive_max}
\alias{recursive_min}
\title{Standard functions implemented in a functional style}
\usage{
recursive_add(x)

recursive_multiply(x)

recursive_max(x)

recursive_min(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
Standard functions implemented in a functional style
}
\details{
Functions leverage either fold right or fold left.
}
\examples{
library(functionalPlayground)

x <- 1:10
recursive_add(x)
recursive_multiply(x)
recursive_max(x)
recursive_min(x)
}
