% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_fold_logical_functions.R,
%   R/8_function_operators.R
\name{recursive_all}
\alias{recursive_all}
\alias{recursive_any}
\alias{recursive_none}
\title{Standard logical functions implemented by folding}
\usage{
recursive_all(x)

recursive_any(x)

recursive_none(x)
}
\arguments{
\item{x}{A logical vector.}
}
\description{
Standard logical functions implemented by folding
}
\details{
Functions leveraging either fold right or fold left.
}
\examples{
library(functionalPlayground)

x <- c(TRUE, TRUE, FALSE, FALSE)
recursive_all(x)
recursive_any(x)
recursive_none(x)
}
