% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4 fold logical functions.R
\name{recursive_all}
\alias{recursive_all}
\alias{recursive_any}
\title{Standard functions implemented in a functional style}
\usage{
recursive_all(x)

recursive_any(x)
}
\arguments{
\item{x}{A logical vector.}
}
\description{
Standard functions implemented in a functional style
}
\details{
Functions leverage either fold right or fold left.
}
\examples{
library(functionalPlayground)

x <- c(TRUE, TRUE, FALSE, FALSE)
recursive_all(x)
recursive_any(x)
}
