% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_function_operators.R
\name{recursive_filter}
\alias{recursive_filter}
\title{Standard function operator.}
\usage{
recursive_filter(f, x)
}
\arguments{
\item{f}{A function that returns TRUE or FALSE.}

\item{x}{A vector to be filtered.}
}
\description{
Standard function operator.
}
\details{
Filter any function.
}
\examples{
library(functionalPlayground)

always_true <- function(...) {
  return(TRUE)
}
always_false <- negate(always_true)
recursive_filter(always_true, 1:10)
recursive_filter(always_false, 1:10)
}
