% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8_function_operators.R
\name{recursive_filter}
\alias{recursive_filter}
\alias{first}
\alias{last}
\title{Standard function operator.}
\usage{
recursive_filter(f, x)

first(f, x)

last(f, x)
}
\arguments{
\item{f}{A function that returns TRUE or FALSE.}

\item{x}{A vector to be filtered.}
}
\description{
Standard function operator.
}
\details{
Filter any function.
}
\examples{
library(functionalPlayground)

recursive_filter(is_even, 1:10)
recursive_filter(is_odd, 1:10)

first(is_even, 1:10)
first(is_odd, 1:10)

last(is_even, 1:10)
last(is_odd, 1:10)
}
