% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_fold_logical_functions.R
\name{recursive_is_in}
\alias{recursive_is_in}
\title{Standard logical functions implemented by mapping and folding.}
\usage{
recursive_is_in(x, y)
}
\arguments{
\item{x}{A vector.}

\item{y}{A vector to check against. Is x in y?}
}
\description{
Standard logical functions implemented by mapping and folding.
}
\details{
Functions leveraging either fold right or fold left and mapping.
}
\examples{
library(functionalPlayground)

recursive_is_in(c(1, 3, 5), c(1, 4))
}
