% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_basic_recursive_functions.R
\name{recursive_reverse}
\alias{recursive_reverse}
\title{Standard functions implemented by recursion}
\usage{
recursive_reverse(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Standard functions implemented by recursion
}
\details{
Functions leveraging recursion.
}
\examples{
library(functionalPlayground)

x <- 1:10
recursive_reverse(x)
}
