% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/9_basic_recursive_functions.R
\name{recursive_sort}
\alias{recursive_sort}
\title{Standard functions implemented by recursion}
\usage{
recursive_sort(x, decreasing = FALSE)
}
\arguments{
\item{x}{A vector that is not empty.}

\item{decreasing}{Logical. Should the vector be sorted in descending order?}
}
\description{
Standard functions implemented by recursion
}
\details{
Functions leveraging recursion.
}
\examples{
library(functionalPlayground)

recursive_sort(1:10)
recursive_sort(10:1, TRUE)
}
